api/requirements.txt
Django==2.0.13
djangorestframework==3.7.7
psycopg2==2.7.3.2
pytz==2017.3

